+++ Car/LKW.java
@@ -3,7 +3,7 @@ public class LKW extends Fahrzeug {
     private int maxAnhaenger;
 
     // Konstruktor
-    public LKW(float tankvolumen, float verbrauch, int maxAnhaenger) {
+    public LKW(double tankvolumen, double verbrauch, int maxAnhaenger) {
         super(tankvolumen, verbrauch);
         this.anzahlAnhaenger = 0;
         this.maxAnhaenger = maxAnhaenger;
